Try to suppress warnings but not work? Do this later...

+++ CoconutEngine/src/Coconut/GameMap/MapTweaker.cpp
@@ -48,7 +48,7 @@ void Coconut::MapTweaker::fillZeros(std::vector<std::vector<int>>& map) {
 
 
 void Coconut::MapTweaker::fitSize(std::vector<std::vector<int>>& map) {
-	m_row = map.size();
+	m_row = static_cast<int> (map.size());
 	m_column = (m_row > 0) ? map[0].size() : 0;
 }
